refactor: update import paths and enhance icon usage in layout components
